Resolve icon user profile, and some owner view
